+++ src/events/MessageCreate.ts
@@ -1,8 +1,7 @@
 import type { ArgsOf, Client } from "discordx";
 import { Discord, On } from "discordx";
-import { addMessageDb } from "../modules/messages/addMessageDb.js";
+import { MessagesModule } from "../modules/messages/Messages.module.js";
 import { checkWarnings } from "../modules/messages/checkWarnings.js";
-import { cleanUpVerifyChannel } from "../modules/messages/cleanUpVerifyChannel.js";
 import { translateReply } from "../modules/messages/translateReply.js";
 import { replyChatGPT } from "../utils/chatgpt/replyChatGPT.js";
 
@@ -11,7 +10,7 @@ export class MessageCreate {
   @On()
   async messageCreate([message]: ArgsOf<"messageCreate">, client: Client) {
     // remove regular messages in verify channel
-    await cleanUpVerifyChannel(message);
+    await MessagesModule.cleanUpVerifyChannel(message);
 
     //delete messages with links
     await checkWarnings(message);
@@ -23,6 +22,6 @@ export class MessageCreate {
     await translateReply(message);
 
     // add Message to Database for statistics
-    await addMessageDb(message);
+    await MessagesModule.addMessageDb(message);
   }
 }
