+++ src/events/MessageDelete.ts
@@ -1,11 +1,11 @@
 import type { ArgsOf, Client } from "discordx";
 import { Discord, On } from "discordx";
-import { deleteMessageDb } from "../modules/messages/deleteMessageDb.js";
+import { MessagesModule } from "../modules/messages/Messages.module.js";
 
 @Discord()
 export class MessageDelete {
   @On()
   async messageDelete([message]: ArgsOf<"messageDelete">, client: Client) {
-    await deleteMessageDb(message);
+    await MessagesModule.deleteMessageDb(message);
   }
 }
